refactor: dto패키지 db에서 분리, MIME_TYPES 전역 상수로 변경
