+++ tripeer/src/main/java/com/j10d207/tripeer/history/controller/HistoryController.java
@@ -15,12 +15,12 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.j10d207.tripeer.history.db.dto.CostReqDTO;
-import com.j10d207.tripeer.history.db.dto.CostResDTO;
-import com.j10d207.tripeer.history.db.dto.GalleryDTO;
-import com.j10d207.tripeer.history.db.dto.GalleryIdListDTO;
-import com.j10d207.tripeer.history.db.dto.HistoryDetailResDTO;
-import com.j10d207.tripeer.history.db.dto.PlanSaveReqDTO;
+import com.j10d207.tripeer.history.dto.request.CostReqDTO;
+import com.j10d207.tripeer.history.dto.response.CostResDTO;
+import com.j10d207.tripeer.history.dto.response.GalleryDTO;
+import com.j10d207.tripeer.history.dto.request.GalleryIdListDTO;
+import com.j10d207.tripeer.history.dto.response.HistoryDetailResDTO;
+import com.j10d207.tripeer.history.dto.request.PlanSaveReqDTO;
 import com.j10d207.tripeer.history.service.GalleryService;
 import com.j10d207.tripeer.history.service.HistoryService;
 import com.j10d207.tripeer.plan.db.dto.PlanListResDTO;
