+++ tripeer/src/main/java/com/j10d207/tripeer/history/dto/response/HistorySpotResDTO.java
@@ -1,0 +1,17 @@
+package com.j10d207.tripeer.history.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class HistorySpotResDTO {
+	private long planDetailId;
+	private String title;
+	private String contentType;
+	private String address;
+	private String image;
+	private int day;
+	private int step;
+	private int cost;
+}
